# docs/api/openapi-v1.yaml
openapi: 3.1.0
info:
  title: Finanzas App API
  version: 1.0.0
  description: API v1 para app financiera Chile-first (CLP).
servers:
  - url: https://api.finanzasapp.cl/v1
security:
  - bearerAuth: []
...
paths:
...
  /transactions:
    get:
      tags: [Transactions]
      summary: Lista movimientos con filtros
      parameters:
        - in: query
          name: month
          schema:
            type: string
            pattern: '^\d{4}-\d{2}$'
        - in: query
          name: type
          schema:
            type: string
            enum: [income, expense]
        - in: query
          name: categoryId
          schema:
            type: string
            format: uuid
        - in: query
          name: source
          schema:
            type: string
            enum: [gmail, manual]
      responses:
        '200':
          description: Transacciones
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Transaction'
    post:
      tags: [Transactions]
      summary: Crear movimiento manual
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTransactionRequest'
      responses:
        '201':
          description: Creado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Transaction'
...
  /categories:
    get:
      tags: [Categories]
      summary: Listar categorias
      responses:
        '200':
          description: Categorias
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Category'
    post:
      tags: [Categories]
      summary: Crear categoria
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
                icon:
                  type: string
                colorHex:
                  type: string
      responses:
        '201':
          description: Creada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'
  /goals:
    get:
      tags: [Goals]
      summary: Listar metas
      responses:
        '200':
          description: Metas
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/SavingGoal'
    post:
      tags: [Goals]
      summary: Crear meta
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateGoalRequest'
      responses:
        '201':
          description: Meta creada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SavingGoal'
...
  /budgets:
    get:
      tags: [Budgets]
      summary: Listar presupuestos mensuales
      parameters:
        - in: query
          name: month
          required: true
          schema:
            type: string
            pattern: '^\d{4}-\d{2}$'
      responses:
        '200':
          description: Presupuestos
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Budget'
    put:
      tags: [Budgets]
      summary: Upsert de presupuesto por categoria y mes
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpsertBudgetRequest'
      responses:
        '200':
          description: Presupuesto guardado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Budget'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Category:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        icon: { type: string }
        colorHex: { type: string }
    Transaction:
      type: object
      properties:
        id: { type: string, format: uuid }
        type: { type: string, enum: [income, expense] }
        source: { type: string, enum: [gmail, manual] }
        amountClp: { type: integer }
        occurredAt: { type: string, format: date-time }
        rawGlosa: { type: string }
        merchant: { type: string }
        categoryId: { type: string, format: uuid }
    CreateTransactionRequest:
      type: object
      required: [type, amountClp, occurredAt, rawGlosa]
      properties:
        type: { type: string, enum: [income, expense] }
        amountClp: { type: integer, minimum: 1 }
        occurredAt: { type: string, format: date-time }
        rawGlosa: { type: string }
        categoryId: { type: string, format: uuid }
    CreateGoalRequest:
      type: object
      required: [name, targetAmountClp, monthlyContributionClp]
      properties:
        name: { type: string }
        targetAmountClp: { type: integer, minimum: 1 }
        monthlyContributionClp: { type: integer, minimum: 1 }
        targetDate: { type: string, format: date }
    SavingGoal:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        targetAmountClp: { type: integer }
        monthlyContributionClp: { type: integer }
        currentAmountClp: { type: integer }
        status: { type: string, enum: [active, paused, completed, cancelled] }
    UpsertBudgetRequest:
      type: object
      required: [categoryId, year, month, amountLimitClp]
      properties:
        categoryId: { type: string, format: uuid }
        year: { type: integer, minimum: 2020 }
        month: { type: integer, minimum: 1, maximum: 12 }
        amountLimitClp: { type: integer, minimum: 1 }
    Budget:
      type: object
      properties:
        id: { type: string, format: uuid }
        categoryId: { type: string, format: uuid }
        year: { type: integer }
        month: { type: integer }
        amountLimitClp: { type: integer }
...